check whether a point lies in the given ellipse
So given a test point (x,y). If the inequality is satisfied, then it is
inside the ellipse; otherwise it is outside the ellipse. Moreover, the
point is on the boundary of the region (i.e., on the ellipse) if and
only if the inequality is satisfied tightly (i.e., the left hand side
evaluates to 1).